feat(color): add scaled HSBK accessors and replace()
Add pre-scaled views of the HSBK components for consumers that do not want
the internal 0.0-1.0 fractions:

- `saturation_pct` / `brightness_pct` return unrounded percentages (0.0-100.0)
- `brightness_uint8` returns an 8-bit value (0-255)

Home Assistant wants `(hue, saturation_pct)` for a light's `hs_color` and
`brightness_uint8` for its `brightness`, so no rescaling is needed at the
integration layer. `brightness_uint8` rounds to nearest rather than
truncating: rounding halves the worst-case error, keeps it unbiased, and
tolerates the float drift accumulated by `lerp_hsb`, `lerp_oklab` and
`average` that would otherwise break the uint8 round-trip.

Add `replace()` to override an arbitrary subset of components in one call
rather than chaining the single-component `with_*` methods. It is named for
`dataclasses.replace`; a blend-style name would collide conceptually with
`lerp_hsb` / `lerp_oklab` / `average`, which are the methods that actually
combine two colours.

Also fix hue quantisation in `to_protocol()`. The expression rounded before
dividing and then truncated the quotient:

    int(round(0x10000 * hue) / 360) % 0x10000

which biased arbitrary hues low by up to 1 LSB. It now rounds the quotient
instead. Values originating from `from_protocol()` are unaffected — all 65536
uint16 hues round-trip identically under both forms. `aurora.py` carried a
copy of the same expression in its direct-to-protocol path; it is updated to
match, which restores agreement between `generate_frame()` and
`generate_protocol_frame()`.

The redundant `int()` wrappers around `round()` in both files are dropped;
bare `round()` already returns an int.

+++ src/lifx/effects/aurora.py
@@ -243,9 +243,9 @@ class EffectAurora(FrameEffect):
             saturation = 0.7 + 0.3 * sin(position * 2 * pi)
 
             # Convert directly to uint16 protocol values
-            hue_u16 = int(round(0x10000 * hue) / 360) % 0x10000
-            sat_u16 = int(round(0xFFFF * saturation))
-            bri_u16 = int(round(0xFFFF * pixel_brightness))
+            hue_u16 = round(0x10000 * hue / 360) % 0x10000
+            sat_u16 = round(0xFFFF * saturation)
+            bri_u16 = round(0xFFFF * pixel_brightness)
 
             result.append((hue_u16, sat_u16, bri_u16, KELVIN_NEUTRAL))
